XP-1265 Rendering without Template - Reload controller dropdown after changes to Site config
-Also refactored controller dropdowns
