+++ modules/admin-ui/src/main/resources/web/admin/apps/content-manager/js/app/wizard/page/contextwindow/inspect/page/PageControllerSelector.ts
@@ -2,59 +2,24 @@ module app.wizard.page.contextwindow.inspect.page {
 
     import PropertyChangedEvent = api.PropertyChangedEvent;
     import LiveEditModel = api.liveedit.LiveEditModel;
-    import SiteModel = api.content.site.SiteModel;
     import PageModel = api.content.page.PageModel;
-    import SetController = api.content.page.SetController;
     import PageDescriptor = api.content.page.PageDescriptor;
     import DescriptorKey = api.content.page.DescriptorKey;
-    import PageDescriptorsJson = api.content.page.PageDescriptorsJson;
-    import PageDescriptorDropdown = api.content.page.PageDescriptorDropdown;
-    import GetPageDescriptorsByApplicationsRequest = api.content.page.GetPageDescriptorsByApplicationsRequest;
-    import OptionSelectedEvent = api.ui.selector.OptionSelectedEvent;
     import LoadedDataEvent = api.util.loader.event.LoadedDataEvent;
 
-    export class PageControllerSelector extends PageDescriptorDropdown {
-
-        private getPageDescriptorsByApplicationsRequest: GetPageDescriptorsByApplicationsRequest;
+    export class PageControllerSelector extends api.content.page.PageDescriptorDropdown {
 
         private pageModel: PageModel;
 
-        private siteModel: SiteModel;
+        constructor(model: LiveEditModel) {
+            super(model);
 
-        constructor() {
-
-            this.getPageDescriptorsByApplicationsRequest = new GetPageDescriptorsByApplicationsRequest([]);
-
-            super('page-controller', {
-                loader: new api.util.loader.BaseLoader<PageDescriptorsJson, PageDescriptor>(this.getPageDescriptorsByApplicationsRequest).
-                    setComparator(new api.content.page.DescriptorByDisplayNameComparator())
-            });
-
-            this.onOptionSelected((event: OptionSelectedEvent<PageDescriptor>) => {
-                var pageDescriptor = event.getOption().displayValue;
-                var setController = new SetController(this).setDescriptor(pageDescriptor);
-                this.pageModel.setController(setController);
-            });
-        }
-
-        setModel(model: LiveEditModel) {
-
-            this.siteModel = model.getSiteModel();
             this.pageModel = model.getPageModel();
 
-            this.getPageDescriptorsByApplicationsRequest.setApplicationKeys(this.siteModel.getApplicationKeys());
             this.onLoadedData((event: LoadedDataEvent<PageDescriptor>) => {
 
                 if (this.pageModel.hasController()) {
                     this.selectController(this.pageModel.getController().getKey());
-                }
-            });
-            this.load();
-
-            this.siteModel.onPropertyChanged((event: api.PropertyChangedEvent) => {
-                if (event.getPropertyName() == SiteModel.PROPERTY_NAME_SITE_CONFIGS) {
-                    this.getPageDescriptorsByApplicationsRequest.setApplicationKeys(this.siteModel.getApplicationKeys());
-                    this.load();
                 }
             });
