--- modules/admin-ui/src/main/resources/web/admin/common/js/content/page/inputtype/pagecontroller/_module.ts
@@ -1,1 +1,0 @@
-///<reference path='PageController.ts' />
