+++ modules/admin-ui/src/main/resources/web/admin/common/js/liveedit/PagePlaceholder.ts
@@ -4,7 +4,6 @@ module api.liveedit {
     import PageDescriptor = api.content.page.PageDescriptor;
     import SetController = api.content.page.SetController;
     import SiteModel = api.content.site.SiteModel;
-    import PageController = api.content.page.inputtype.pagecontroller.PageController;
     import PageDescriptorDropdown = api.content.page.PageDescriptorDropdown;
     import PageDescriptorsJson = api.content.page.PageDescriptorsJson;
     import OptionSelectedEvent = api.ui.selector.OptionSelectedEvent;
@@ -22,21 +21,8 @@ module api.liveedit {
             this.addClass("page-placeholder");
             this.pageView = pageView;
 
-            var applicationKeys = pageView.liveEditModel.getSiteModel().getApplicationKeys(),
-                request = new GetPageDescriptorsByApplicationsRequest(applicationKeys),
-                loader = new api.util.loader.BaseLoader<PageDescriptorsJson, PageDescriptor>(request);
-
-            this.controllerDropdown = new PageDescriptorDropdown('page-controller', {
-                loader: loader
-            });
+            this.controllerDropdown = new PageDescriptorDropdown(pageView.liveEditModel);
             this.controllerDropdown.load();
-
-            this.controllerDropdown.onOptionSelected((event: api.ui.selector.OptionSelectedEvent<api.content.page.PageDescriptor>) => {
-
-                var setController = new SetController(this).
-                    setDescriptor(event.getOption().displayValue);
-                this.pageView.liveEditModel.getPageModel().setController(setController);
-            });
 
             this.appendChild(this.controllerDropdown);
 
@@ -50,12 +36,6 @@ module api.liveedit {
                 event.stopPropagation();
             });
 
-            pageView.liveEditModel.getSiteModel().onPropertyChanged((event: api.PropertyChangedEvent) => {
-                if (event.getPropertyName() == SiteModel.PROPERTY_NAME_SITE_CONFIGS) {
-                    request.setApplicationKeys(pageView.liveEditModel.getSiteModel().getApplicationKeys());
-                    loader.load();
-                }
-            });
         }
 
         select() {
